patch warning

+++ src/map/load_layer.c
@@ -25,7 +25,7 @@ layer_t **load_layer(const char *pathfile, map_object_t **map_object)
     char *line = NULL;
     size_t len = 0;
     int index = 0;
-    size_t read_size = 0;
+    int read_size = 0;
     while ((read_size = getline(&line, &len, fp)) != -1) {
         line[read_size - 1] = '\0';
         if (my_strncmp(line, "layer", 5) == 0) {
